+++ modules/project/ui/src/sap/commerce/toolset/project/view/nodes/AiAgentPsiDirectoryNode.kt
@@ -1,0 +1,45 @@
+/*
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package sap.commerce.toolset.project.view.nodes
+
+import com.intellij.ide.projectView.PresentationData
+import com.intellij.ide.projectView.ViewSettings
+import com.intellij.ide.projectView.impl.nodes.PsiDirectoryNode
+import com.intellij.openapi.project.Project
+import com.intellij.psi.PsiDirectory
+import com.intellij.ui.SimpleTextAttributes
+import javax.swing.Icon
+
+class AiAgentPsiDirectoryNode(
+    project: Project,
+    value: PsiDirectory,
+    viewSettings: ViewSettings,
+    private val icon: Icon
+) : PsiDirectoryNode(project, value, viewSettings) {
+
+    override fun update(presentation: PresentationData) = with(presentation) {
+        setIcon(icon)
+        addText(
+            ColoredFragment(
+                value.name,
+                SimpleTextAttributes.REGULAR_ATTRIBUTES
+            )
+        )
+    }
+}